Moves genUniqueArr helper function into data/helpers.js
