+++ data/seeds/001_users.js
@@ -1,15 +1,5 @@
 const faker = require("faker");
-
-const genUniqueArr = (num, fakerFn) => {
-  let arr = [];
-  while (arr.length < num) {
-    const temp = fakerFn();
-    if (!arr.includes(temp)) {
-      arr.push(temp);
-    }
-  }
-  return arr;
-};
+const {genUniqueArr} = require("../helpers.js");
 
 const generateSeeds = () => {
   const numOfUsers = 500;
